refactor: move probe ready conditional to doProbe function
